reimplement a template version of udp_server
