+++ util/CMakeLists.txt
@@ -76,7 +76,6 @@ if(UNIX)
   list(APPEND CUB_MASTER_SOURCES ${MASTER_DIR}/heartbeat_cluster.cpp)
   list(APPEND CUB_MASTER_SOURCES ${MASTER_DIR}/heartbeat_service.cpp)
   list(APPEND CUB_MASTER_SOURCES ${MASTER_DIR}/master_heartbeat.cpp)
-  list(APPEND CUB_MASTER_SOURCES ${MASTER_DIR}/udp_rpc.cpp)
   SET_SOURCE_FILES_PROPERTIES(
 	${CUB_MASTER_SOURCES}
 	PROPERTIES LANGUAGE CXX
